[Refactor] PlayerStats 합치기 진행
따로 떨어진 PlayerStats를 모으는 작업 진행
